+++ tests/ast/tastop.nim
@@ -75,31 +75,31 @@ endmodule
       port: new_node(NkPortDecl, li(2, 4), @[
          new_identifier_node(NkDirection, li(2, 4), "input"),
          new_identifier_node(NkNetType, li(2, 10), "wire"),
-         new_identifier_node(NkPortIdentifier, li(2, 15), "clk_i"),
-         new_identifier_node(NkPortIdentifier, li(3, 15), "another_port_i"),
-         new_identifier_node(NkPortIdentifier, li(4, 15), "the_last_port_i")
+         new_identifier_node(NkIdentifier, li(2, 15), "clk_i"),
+         new_identifier_node(NkIdentifier, li(3, 15), "another_port_i"),
+         new_identifier_node(NkIdentifier, li(4, 15), "the_last_port_i")
       ]),
-      identifier: new_identifier_node(NkPortIdentifier, li(2, 15), "clk_i")
+      identifier: new_identifier_node(NkIdentifier, li(2, 15), "clk_i")
    ),
    (
       port: new_node(NkPortDecl, li(2, 4), @[
          new_identifier_node(NkDirection, li(2, 4), "input"),
          new_identifier_node(NkNetType, li(2, 10), "wire"),
-         new_identifier_node(NkPortIdentifier, li(2, 15), "clk_i"),
-         new_identifier_node(NkPortIdentifier, li(3, 15), "another_port_i"),
-         new_identifier_node(NkPortIdentifier, li(4, 15), "the_last_port_i")
+         new_identifier_node(NkIdentifier, li(2, 15), "clk_i"),
+         new_identifier_node(NkIdentifier, li(3, 15), "another_port_i"),
+         new_identifier_node(NkIdentifier, li(4, 15), "the_last_port_i")
       ]),
-      identifier: new_identifier_node(NkPortIdentifier, li(3, 15), "another_port_i")
+      identifier: new_identifier_node(NkIdentifier, li(3, 15), "another_port_i")
    ),
    (
       port: new_node(NkPortDecl, li(2, 4), @[
          new_identifier_node(NkDirection, li(2, 4), "input"),
          new_identifier_node(NkNetType, li(2, 10), "wire"),
-         new_identifier_node(NkPortIdentifier, li(2, 15), "clk_i"),
-         new_identifier_node(NkPortIdentifier, li(3, 15), "another_port_i"),
-         new_identifier_node(NkPortIdentifier, li(4, 15), "the_last_port_i"),
+         new_identifier_node(NkIdentifier, li(2, 15), "clk_i"),
+         new_identifier_node(NkIdentifier, li(3, 15), "another_port_i"),
+         new_identifier_node(NkIdentifier, li(4, 15), "the_last_port_i"),
       ]),
-      identifier: new_identifier_node(NkPortIdentifier, li(4, 15), "the_last_port_i")
+      identifier: new_identifier_node(NkIdentifier, li(4, 15), "the_last_port_i")
    ),
    (
       port: new_node(NkPortDecl, li(5, 4), @[
@@ -109,9 +109,9 @@ endmodule
             new_inumber_node(NkIntLit, li(5, 16), "7", Base10, -1),
             new_inumber_node(NkIntLit, li(5, 18), "0", Base10, -1)
          ]),
-         new_identifier_node(NkPortIdentifier, li(5, 21), "data_o")
+         new_identifier_node(NkIdentifier, li(5, 21), "data_o")
       ]),
-      identifier: new_identifier_node(NkPortIdentifier, li(5, 21), "data_o")
+      identifier: new_identifier_node(NkIdentifier, li(5, 21), "data_o")
    )
 ])
 
@@ -130,31 +130,31 @@ endmodule
       port: new_node(NkPortDecl, li(4, 4), @[
          new_identifier_node(NkDirection, li(4, 4), "input"),
          new_identifier_node(NkNetType, li(4, 10), "wire"),
-         new_identifier_node(NkPortIdentifier, li(4, 15), "clk_i"),
-         new_identifier_node(NkPortIdentifier, li(5, 15), "another_port_i"),
-         new_identifier_node(NkPortIdentifier, li(6, 15), "the_last_port_i")
+         new_identifier_node(NkIdentifier, li(4, 15), "clk_i"),
+         new_identifier_node(NkIdentifier, li(5, 15), "another_port_i"),
+         new_identifier_node(NkIdentifier, li(6, 15), "the_last_port_i")
       ]),
-      identifier: new_identifier_node(NkPortIdentifier, li(4, 15), "clk_i")
+      identifier: new_identifier_node(NkIdentifier, li(4, 15), "clk_i")
    ),
    (
       port: new_node(NkPortDecl, li(4, 4), @[
          new_identifier_node(NkDirection, li(4, 4), "input"),
          new_identifier_node(NkNetType, li(4, 10), "wire"),
-         new_identifier_node(NkPortIdentifier, li(4, 15), "clk_i"),
-         new_identifier_node(NkPortIdentifier, li(5, 15), "another_port_i"),
-         new_identifier_node(NkPortIdentifier, li(6, 15), "the_last_port_i")
+         new_identifier_node(NkIdentifier, li(4, 15), "clk_i"),
+         new_identifier_node(NkIdentifier, li(5, 15), "another_port_i"),
+         new_identifier_node(NkIdentifier, li(6, 15), "the_last_port_i")
       ]),
-      identifier: new_identifier_node(NkPortIdentifier, li(5, 15), "another_port_i")
+      identifier: new_identifier_node(NkIdentifier, li(5, 15), "another_port_i")
    ),
    (
       port: new_node(NkPortDecl, li(4, 4), @[
          new_identifier_node(NkDirection, li(4, 4), "input"),
          new_identifier_node(NkNetType, li(4, 10), "wire"),
-         new_identifier_node(NkPortIdentifier, li(4, 15), "clk_i"),
-         new_identifier_node(NkPortIdentifier, li(5, 15), "another_port_i"),
-         new_identifier_node(NkPortIdentifier, li(6, 15), "the_last_port_i")
+         new_identifier_node(NkIdentifier, li(4, 15), "clk_i"),
+         new_identifier_node(NkIdentifier, li(5, 15), "another_port_i"),
+         new_identifier_node(NkIdentifier, li(6, 15), "the_last_port_i")
       ]),
-      identifier: new_identifier_node(NkPortIdentifier, li(6, 15), "the_last_port_i")
+      identifier: new_identifier_node(NkIdentifier, li(6, 15), "the_last_port_i")
    ),
    (
       port: new_node(NkPortDecl, li(7, 4), @[
@@ -164,9 +164,9 @@ endmodule
             new_inumber_node(NkIntLit, li(7, 16), "7", Base10, -1),
             new_inumber_node(NkIntLit, li(7, 18), "0", Base10, -1)
          ]),
-         new_identifier_node(NkPortIdentifier, li(7, 21), "data_o")
+         new_identifier_node(NkIdentifier, li(7, 21), "data_o")
       ]),
-      identifier: new_identifier_node(NkPortIdentifier, li(7, 21), "data_o")
+      identifier: new_identifier_node(NkIdentifier, li(7, 21), "data_o")
    )
 ])
 
@@ -182,20 +182,20 @@ endmodule
    (
       parameter: new_node(NkParameterDecl, li(2, 4), @[
          new_node(NkParamAssignment, li(2, 14), @[
-            new_identifier_node(NkParameterIdentifier, li(2, 14), "WIDTH"),
+            new_identifier_node(NkIdentifier, li(2, 14), "WIDTH"),
             new_inumber_node(NkIntLit, li(2, 22), "32", Base10, -1)
          ])
       ]),
-      identifier: new_identifier_node(NkParameterIdentifier, li(2, 14), "WIDTH")
+      identifier: new_identifier_node(NkIdentifier, li(2, 14), "WIDTH")
    ),
    (
       parameter: new_node(NkParameterDecl, li(3, 4), @[
          new_node(NkParamAssignment, li(3, 14), @[
-            new_identifier_node(NkParameterIdentifier, li(3, 14), "ZERO"),
+            new_identifier_node(NkIdentifier, li(3, 14), "ZERO"),
             new_inumber_node(NkIntLit, li(3, 21), "0", Base10, -1)
          ])
       ]),
-      identifier: new_identifier_node(NkParameterIdentifier, li(3, 14), "ZERO")
+      identifier: new_identifier_node(NkIdentifier, li(3, 14), "ZERO")
    )
 ])
 
@@ -208,11 +208,11 @@ endmodule
    (
       parameter: new_node(NkParameterDecl, li(2, 4), @[
          new_node(NkParamAssignment, li(2, 14), @[
-            new_identifier_node(NkParameterIdentifier, li(2, 14), "FOO"),
+            new_identifier_node(NkIdentifier, li(2, 14), "FOO"),
             new_inumber_node(NkIntLit, li(2, 20), "3", Base10, -1)
          ])
       ]),
-      identifier: new_identifier_node(NkParameterIdentifier, li(2, 14), "FOO")
+      identifier: new_identifier_node(NkIdentifier, li(2, 14), "FOO")
    )
 ])
 
@@ -238,9 +238,9 @@ endmodule
       driver: new_node(NkPortDecl, li(2, 4), @[
          new_identifier_node(NkDirection, li(2, 4), "input"),
          new_identifier_node(NkNetType, li(2, 10), "wire"),
-         new_identifier_node(NkPortIdentifier, li(2, 15), "clk_i"),
+         new_identifier_node(NkIdentifier, li(2, 15), "clk_i"),
       ]),
-      identifier: new_identifier_node(NkPortIdentifier, li(2, 15), "clk_i")
+      identifier: new_identifier_node(NkIdentifier, li(2, 15), "clk_i")
    ),
    (
       driver: new_node(NkContinuousAssignment, li(6, 4), @[